+++ core/src/main/java/com/aminovic/core/domain/preferences/Preferences.kt
@@ -13,9 +13,9 @@ interface Preferences {
     fun saveHeight(height: Int)
     fun saveActivityLevel(activityLevel: ActivityLevel)
     fun saveGoalType(goalType: GoalType)
-    fun saveCarbRatio(ratio: Float)
-    fun saveProteinRatio(ratio: Float)
-    fun saveFatRatio(ratio: Float)
+    fun saveCarbRatio(ratio: Int)
+    fun saveProteinRatio(ratio: Int)
+    fun saveFatRatio(ratio: Int)
 
     fun loadUserInfo(): UserInfo
